Add actions list on warn list command

+++ db/WarnAction.py
@@ -1,14 +1,14 @@
 from datetime import timedelta
 
 from db import Base
-from sqlalchemy import Column, Integer, BigInteger, Float, String
+from sqlalchemy import Column, Integer, BigInteger, String
 
 
 class WarnAction(Base):
     __tablename__ = "warn_actions"
     id = Column(Integer, primary_key=True)
     guild = Column(BigInteger, nullable=False)
-    count = Column(Float, nullable=False, unique=True)
+    count = Column(Integer, nullable=False, unique=True)
     action = Column(String, nullable=False)
     duration = Column(BigInteger)
